requirements.txt
asgiref==3.3.1
Django==3.0
django-bootstrap-static==4.2.1
djangorestframework==3.11.2
pkg-resources==0.0.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
